requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cffi==1.14.2
cryptography==2.9.2
Django==3.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
jwt==1.0.0
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
